Update 5 files
-Move AppContents into SeqeuncerFrame component
-Added SelectStrip function and passing it to the Strip control component
